refactor: normalize module naming conventions
- Rename debts.py → debt.py (singular)
- Rename goals.py → goal.py (singular)
- Rename help.py → help_handler.py (consistent suffix)
- Rename GoalsRepository → GoalRepository (singular)
- Update all imports accordingly
