+++ bot/database/__init__.py
@@ -2,9 +2,9 @@ from bot.database.db import DataBase
 from bot.database.repositories.balance import BalanceRepository
 from bot.database.repositories.category import CategoryRepository
 from bot.database.repositories.category_types import CategoryTypeRepository
-from bot.database.repositories.debts import DebtRepository
+from bot.database.repositories.debt import DebtRepository
 from bot.database.repositories.expenses import ExpensesRepository
-from bot.database.repositories.goals import GoalsRepository
+from bot.database.repositories.goal import GoalRepository
 from bot.database.repositories.incomes import IncomesRepository
 
 
@@ -16,4 +16,4 @@ def init_db():
     ExpensesRepository(database).init_table()
     IncomesRepository(database).init_table()
     DebtRepository(database).init_table()
-    GoalsRepository(database).init_table()
+    GoalRepository(database).init_table()
